Authentication and Authorization

+++ src/main/java/com/ayush/blogproject/service/BlogServiceImp.java
@@ -4,16 +4,20 @@ import com.ayush.blogproject.dto.PostBlogDto;
 import com.ayush.blogproject.model.Comments;
 import com.ayush.blogproject.model.Posts;
 import com.ayush.blogproject.model.Tags;
+import com.ayush.blogproject.model.User;
 import com.ayush.blogproject.repository.CommentsRepository;
 import com.ayush.blogproject.repository.PostRepository;
 import com.ayush.blogproject.repository.TagRepository;
+import com.ayush.blogproject.repository.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
+import org.springframework.security.access.AccessDeniedException;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -25,16 +29,25 @@ public class BlogServiceImp implements BlogService {
     PostRepository postRepository;
     TagRepository tagRepository;
     CommentsRepository commentsRepository;
+    UserRepository userRepository;
 
-    public BlogServiceImp(PostRepository postRepository, TagRepository tagRepository, CommentsRepository commentsRepository) {
+    public BlogServiceImp(PostRepository postRepository, TagRepository tagRepository, CommentsRepository commentsRepository,UserRepository userRepository) {
         this.postRepository = postRepository;
         this.tagRepository = tagRepository;
         this.commentsRepository = commentsRepository;
+        this.userRepository=userRepository;
     }
 
     @Override
-    public void savePost(Posts posts, String tagNames, String action) {
+   // public void savePost(Posts posts, String tagNames, String action) {
+    public void savePost(Posts posts, String tagNames, String action, String author) {
         System.out.println("from service: " + action);
+
+        User user = userRepository.findByUsername(author).orElse(null);
+        System.out.println("AUTHOR INPUT: " + author);
+        System.out.println("USER FOUND: " + user);
+        posts.setUser(user);
+
         String tags[] = tagNames.split(",");
         HashSet<Tags> tagSet = new HashSet<>();
         for (String tagName : tags) {
@@ -52,6 +65,8 @@ public class BlogServiceImp implements BlogService {
             }
         }
         posts.setTags(tagSet);
+        posts.setUser(user);
+
         if ("publish".equals(action)) {
             posts.setPublished(true);
             System.out.println("here    trueeeeeeeeeeeeeeeeeeeeeee");
@@ -69,20 +84,6 @@ public class BlogServiceImp implements BlogService {
         postRepository.save(posts);
     }
 
-//        @Override
-//    public Page<Posts> getAllPosts(int page,int limit,String sortBy) {
-//
-//        Sort sort;
-//            if ("publishedAt".equals(sortBy)) {
-//                sort = Sort.by(Sort.Direction.DESC, "publishedAt");
-//            } else {
-//                sort = Sort.by(sortBy).ascending();
-//            }
-//            Pageable pageable = PageRequest.of(page, limit, sort);
-//            return postRepository.findAllPublished(pageable);
-//        }
-
-
     @Override
     public Page<Posts> getAllPosts(int page, int limit, String sortBy, String search, String author, Long tagId) {
 
@@ -92,7 +93,7 @@ public class BlogServiceImp implements BlogService {
         // JPQL ke liye — Java field naam
         Sort jpqlSort = "publishedAt".equals(sortBy)
                 ? Sort.by(Sort.Direction.DESC, "publishedAt")
-                : Sort.by(Sort.Direction.ASC, sortBy.equals("content") ? "content" : sortBy.equals("author") ? "author" : "title");
+                : Sort.by(Sort.Direction.ASC, sortBy.equals("content") ? "content" : sortBy.equals("author") ? "user.username" : "title");
 
         Pageable jpqlPageable = PageRequest.of(page, limit, jpqlSort);
 
@@ -105,7 +106,7 @@ public class BlogServiceImp implements BlogService {
         String nativeSortColumn;
         switch (sortBy) {
             case "publishedAt" -> nativeSortColumn = "published_at";
-            case "author" -> nativeSortColumn = "author";
+            case "author" -> nativeSortColumn = "sort_username";
             case "content" -> nativeSortColumn = "content";
             default -> nativeSortColumn = "title";
         }
@@ -139,6 +140,18 @@ public class BlogServiceImp implements BlogService {
     @Override
     @Transactional
     public Posts updateBlog(Posts posts, String tagNames) {
+
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String userName = auth.getName();
+
+        boolean isOwner = posts.getUser().getUsername().equals(userName);
+        boolean isAdmin = auth.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isOwner && !isAdmin) {
+            throw new AccessDeniedException("Not allowed!");
+        }
+
         Posts existing = postRepository.findById(posts.getId()).orElseThrow();
         String tags[] = tagNames.split(",");
         HashSet<Tags> tagSet = new HashSet<>();
@@ -156,6 +169,7 @@ public class BlogServiceImp implements BlogService {
                 tagSet.add(tag);
             }
         }
+        posts.setUser(existing.getUser());
         posts.setTags(tagSet);
         posts.setPublished(existing.isPublished());
         posts.setPublishedAt(existing.getPublishedAt());
@@ -171,6 +185,18 @@ public class BlogServiceImp implements BlogService {
 
     @Transactional
     public void deleteBlog(Long id) {
+
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String userName = auth.getName();
+        Posts posts=postRepository.findById(id).orElse(null);
+        boolean isOwner = posts.getUser().getUsername().equals(userName);
+        boolean isAdmin = auth.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isOwner && !isAdmin) {
+            throw new AccessDeniedException("Not allowed!");
+        }
+
         postRepository.deleteById(id);
     }
